requirements.txt
asttokens==3.0.0
backcall==0.2.0
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
fsspec==2024.10.0
gym==0.26.2
gym-notices==0.0.8
gym-super-mario-bros==7.4.0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-extensions==1.0.0
nes_py==8.2.1
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyglet==1.5.21
Pygments==2.18.0
python-dateutil==2.9.0.post0
pyzmq==26.2.0
requests==2.32.3
six==1.16.0
stack-data==0.6.3
sympy==1.13.3
tomli==2.2.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==2.2.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.20.2
